=== v2.0 start === changed to a new DB + Role class model changes

+++ src/main/java/spring/controllers/UserController.java
@@ -52,7 +52,7 @@ public class UserController {
 
     @RequestMapping(value = "/edit", method = RequestMethod.GET)
     //тут творится магия! Передавай хоть инт, хоть стринг, будет работать как надо
-    public String toEditForm(@RequestParam(value = "id") int id, Model model) {
+    public String toEditForm(@RequestParam(value = "id") long id, Model model) {
         model.addAttribute("user", service.getUser(id));
         return "user-form";
     }
